requirements.txt
streamlit
google-genai
faiss-cpu
python-dotenv
langchain
langchain-community
beautifulsoup4
requests
numpy
scikit-learn
